% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdsuite-package.R
\docType{package}
\name{ssdsuite-package}
\alias{ssdsuite}
\alias{ssdsuite-package}
\title{ssdsuite: Install and Load Species Sensitivity Distribution Software}
\description{
Species sensitivity distributions are cumulative probability distributions which are fitted to toxicity concentrations for different species as described by Posthuma et al.(2001) <isbn:9781566705783>. The ssdtools package uses Maximum Likelihood to fit distributions such as the gamma, log-logistic, log-normal and log-normal log-normal mixture. Multiple distributions can be averaged using Akaike Information Criteria. Confidence intervals on hazard concentrations and proportions are produced by bootstrapping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/ssdsuite}
  \item \url{https://bcgov.github.io/ssdsuite/}
  \item Report bugs at \url{https://github.com/bcgov/ssdsuite/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Rebecca Fisher \email{R.Fisher@aims.gov.au}
  \item David Fox \email{david.fox@environmetrics.net.au}
  \item Carl Schwarz
}

Other contributors:
\itemize{
  \item Angeline Tillmanns [contributor]
  \item Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID}) [contributor]
  \item Kathleen McTavish [contributor]
  \item Heather Thompson [contributor]
  \item Doug Spry [contributor]
  \item Rick van Dam [contributor]
  \item Graham Batley [contributor]
  \item Ali Azizishirazi [contributor]
  \item Australian Government Department of Climate Change, Energy, the Environment and Water [funder, copyright holder]
}

}
\keyword{internal}
