% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssdsuite_packages}
\alias{ssdsuite_packages}
\title{List all ssdsuite packages}
\usage{
ssdsuite_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include ssdsuite in the list?}
}
\description{
Inspired by \code{tidyverse::tidyverse_packages()}.
}
\examples{
ssdsuite_packages()
}
