% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdsuite-package.R
\docType{package}
\name{ssdsuite-package}
\alias{ssdsuite}
\alias{ssdsuite-package}
\title{ssdsuite: Install and Load Species Sensitivity Distribution Data and Software}
\description{
Installs and loads species sensitivity distribution data and software.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/ssdsuite}
  \item \url{https://poissonconsulting.github.io/ssdsuite/}
  \item Report bugs at \url{https://github.com/poissonconsulting/ssdsuite/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Rebecca Fisher \email{R.Fisher@aims.gov.au}
  \item David Fox \email{david.fox@environmetrics.net.au}
  \item Carl Schwarz
}

Other contributors:
\itemize{
  \item Angeline Tillmanns [contributor]
  \item Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID}) [contributor]
  \item Kathleen McTavish [contributor]
  \item Heather Thompson [contributor]
  \item Doug Spry [contributor]
  \item Rick van Dam [contributor]
  \item Graham Batley [contributor]
  \item Ali Azizishirazi [contributor]
  \item Province of British Columbia [funder, copyright holder]
  \item Environment and Climate Change Canada [funder, copyright holder]
  \item Australian Government Department of Climate Change, Energy, the Environment and Water [funder, copyright holder]
  \item Australian Institute Marine Science [funder, copyright holder]
  \item Poisson Consulting [funder, copyright holder]
}

}
\keyword{internal}
